IntField, DoubleField, LongField and relevant classes refactored and improved, NumberField disables grouping, must be turned on explicitly UiValues, numerical value methods no longer take a format parameter, the text field does all the formatting EntityUiUtil now supplies cloned format instances to NumberField descendants LocalEntityConnection now logs RecordModifiedExceptions correctly
git-svn-id: svn://svn.hafro.is/jminor/trunk@7922 51816742-9f83-410d-85ed-e7bcf7042953

+++ core/src/test/java/org/jminor/swing/common/ui/control/LongValueLinkTest.java
@@ -24,8 +24,8 @@ public class LongValueLinkTest {
   public void testInteger() throws Exception {
     final LongField txtLong = new LongField();
     ValueLinks.longValueLink(txtLong, this, "longValue", evtLongValueChanged, false, true);
-    assertNull("Long value should be null on initialization", txtLong.getInt());
-    setLongValue(2l);
+    assertNull("Long value should be null on initialization", txtLong.getLong());
+    setLongValue(2L);
     assertEquals("Long value should be 2", 2, txtLong.getLong().longValue());
     txtLong.setText("42");
     assertEquals("Long value should be 42", 42, longValue.longValue());
